Fix panic when creating the ConfigMap watch fails
(*K8sConfigMapWatcher).Watch panics when creating the ConfigMap watcher
fails, as it attempts to call "Stop()" on nil (returned by client-go)
when creating the watch returns an error.  This is pretty simply fixed
by only deferring the call once the error check has occurred. Fixing
this issue will cause errors returned from the Kubernetes API to be
properly logged, rather than causing a panic.

I added a test for this behavior, and confirmed that the test causes a
panic without the relevant code change.