Use name and namespace from the request
It's possible for metadata.namespace and metadata.name to be empty in
the objects passed to the webhook. This was causing the ignored
namespaces check to fail and the name and namespace of the pods was
empty in a number of logging statements.

Additionally, the name in the request may also be omitted, in which case
there is unfortunately no way to know the name of the pod.

I've addressed these issues by:
- Moving the namespaces check into the mutate method and using the
  Namespace field from the request to perform the check
- Removing the logging statements from
  `getSidecarConfigurationRequested` in favour of wrapping the errors it
  returns with the required information. The log statement in the mutate
  function can then log this information, along with the Namespace from
  the request.
- Replacing the pod name with an informative placeholder in cases where
  the pod name cannot be known.